requirements.txt
aniso8601==7.0.0
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.2
dateparser==1.2.0
defusedxml==0.8.0rc2
Django==3.2.12
django-jet==1.0.8
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.5
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==3.0.3
idna==3.6
inflection==0.5.1
itypes==1.2.0
jet-bridge-base==1.8.39
Jinja2==3.1.3
MarkupSafe==2.1.5
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.2
pillow==10.2.0
promise==2.3
prompt-toolkit==2.0.9
psutil==5.9.8
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2024.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
Rx==1.6.3
simplejson==3.19.2
six==1.16.0
social-auth-core==4.5.2
SQLAlchemy==1.4.51
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.0
wcwidth==0.2.13
